feat(CleanDIFT): use and SD15UnetFeatureExtractor to wrap UNet2DConditionMOdel
This allows the extraction of intermediate feature hooks.

The current implementation failed to run on collab because it used all
the RAM memory available.

Considering the ressources available, we should rollback to use the
deprecated pretrained backbone of SD 1.5 avaialble and adapt it using
the current implementation as reference and retrain the projection heads
and feature extractors from scratch instead of using the pretrained
weights provided by the paper to make sure that we don't encounter any
architecture issue in training.

Another alternative should be to try using SD 3.5 medium and adapt the
implementation to apply CleanDIFT method. We will try that after
attempting to use SD1.5 first along with the current implementation as
reference.

Signed-off-by: Neil Taison Rigaud <ntsrigaud14@gmail.com>