fix(CleanDIFT): correct `get_features()` method returned features format to `[B, C, H, W]`
This helps ensure that the we obtain a better accuracy and speedup of
CleanDIFT versus the traditional DIFT method as mentionned in the paper.

But currently, the obtained accuracy is still 27x faster than DIFT, but
it should be 50x faster according to the paper. This will be
investigated and fix if possible then evaluated over the whole SPair-71k
dataset.

Signed-off-by: Neil Taison Rigaud <ntsrigaud14@gmail.com>