docs(CleanDIFT): clarification of training results
We have obtained around 30x speed up when comparing CleanDIFT with the
traditional DIFT because the one that we implemented uses just an
ensemble learning and not DDIM inversion which is used in the original
paper for method comparison.

We will also compare the CleanDIFT with the DDIM inversion implemented
in the traditional DIFT to ensure the accuracy and performance of our
implementation.

Signed-off-by: Neil Taison Rigaud <ntsrigaud14@gmail.com>