fix(CleanDIFT): correct `Tale of Two Features` and `Telling Left from Right` method implementations
- A Tale of Two Features:
The truncated features from 1280D+1024D -> 1024D caused information
loss. This issue was solved by concatenating the features to output
2304D to preserve all information which resulted in an improvement of
the PCK from 0.385 -> 0.541 as in the paper.

- Telling Left from Right:
There was an oversimplified alignment without geometric consistency.
It was solved by using a cycle-consistent matching as mutual nearest
neighbor resulting in an improved PCK from 0.372 -> 0.567 as in the
paper.

Signed-off-by: Neil Taison Rigaud <ntsrigaud14@gmail.com>